lab2 15-18 finished

+++ labs/lab2/PinochleGame.cpp
@@ -2,7 +2,7 @@
  * @FilePath: /428cpp/labs/lab2/PinochleGame.cpp
  * @Author: Zhikuan Wei w.zhikuan@wustl.edu
  * @Date: 2022-09-25 14:31:48
- * @LastEditTime: 2022-10-24 16:48:32
+ * @LastEditTime: 2022-10-25 19:20:08
  * @Description: Definition to PinochleGame.h
  *
  */
@@ -15,12 +15,11 @@ const int EIGHT_SAME_RANKS = 8;
 const int FOUR_SAME_RANKS = 4;
 const int DOUBLE_PINOCHLE_CASE = 2;
 const int PINOCHLE_CASE = 1;
-const int SUIT_MASK = (1 << Suit::undefined) - 1;
+const int SUIT_MASK = (1 << static_cast<int>(Suit::undefined)) - 1;
 
-unsigned int PinochleGame::PinochleMeldsPointValue[PINOCHLE_NUM_ITEMS] = {10, 20, 40, 40, 40, 60, 80, 100, 150, 300, 400, 600, 800, 1000, 1500};
+unsigned int PinochleGame::PinochleMeldsPointValue[PINOCHLE_NUM_ITEMS] = { 10, 20, 40, 40, 40, 60, 80, 100, 150, 300, 400, 600, 800, 1000, 1500 };
 
-PinochleGame::PinochleGame(int argc, const char *argv[]) : Game(argc, argv)
-{
+PinochleGame::PinochleGame(int argc, const char* argv[]) : Game(argc, argv) {
     // create as many hands as players in the game
     for (int i = 2; i < argc; ++i)
     {
@@ -29,8 +28,7 @@ PinochleGame::PinochleGame(int argc, const char *argv[]) : Game(argc, argv)
     }
 }
 
-void PinochleGame::deal()
-{
+void PinochleGame::deal() {
     // size of a packet
     const int packet = 3;
     // number of players in game
@@ -39,7 +37,7 @@ void PinochleGame::deal()
     for (int p = 0; !deck.isEmpty(); p = (p + 1) % numHands)
     {
         // current hand
-        auto &hand = hands[p];
+        auto& hand = hands[p];
         // shift a packet of cards
         for (int i = 0; i < packet && !deck.isEmpty(); ++i)
         {
@@ -48,8 +46,7 @@ void PinochleGame::deal()
     }
 }
 
-int PinochleGame::play()
-{
+int PinochleGame::play() {
     const int CardsInRow = 8;
     const int STOP = 0;
 
@@ -83,8 +80,7 @@ int PinochleGame::play()
     }
 }
 
-void PinochleGame::print(std::ostream &os, const std::size_t rc)
-{
+void PinochleGame::print(std::ostream& os, const std::size_t rc) {
     int numPlayer = players.size();
     // print each player's status
     for (int i = 0; i < numPlayer; ++i)
@@ -94,8 +90,7 @@ void PinochleGame::print(std::ostream &os, const std::size_t rc)
     }
 }
 
-void PinochleGame::collectAll()
-{
+void PinochleGame::collectAll() {
     int numPlayer = players.size();
     // collect cards from each player hand
     for (int i = 0; i < numPlayer; ++i)
@@ -104,82 +99,79 @@ void PinochleGame::collectAll()
     }
 }
 
-bool PinochleGame::askForStop(std::ostream &os, std::istream &is)
-{
+bool PinochleGame::askForStop(std::ostream& os, std::istream& is) {
     std::string input;
     // prompt hint message
     os << std::endl
-       << ">>> Do you wish to stop playing? (yes/*) " << std::endl;
+        << ">>> Do you wish to stop playing? (yes/*) " << std::endl;
     // read in a string from the standard input stream
     is >> input;
     // return whether input is yes
     return "yes" == input;
 }
 
-std::ostream &operator<<(std::ostream &os, const PinochleMelds &pm)
-{
+std::ostream& operator<<(std::ostream& os, const PinochleMelds& pm) {
     std::string meldName;
     switch (pm)
     {
-    case PinochleMelds::dix:
-        meldName = "dix";
-        break;
-    case PinochleMelds::offsuitmarriage:
-        meldName = "offsuitmarriage";
-        break;
-    case PinochleMelds::fortyjacks:
-        meldName = "fortyjacks";
-        break;
-    case PinochleMelds::pinochle:
-        meldName = "pinochle";
-        break;
-    case PinochleMelds::insuitmarriage:
-        meldName = "insuitmarriage";
-        break;
-    case PinochleMelds::sixtyqueens:
-        meldName = "sixtyqueens";
-        break;
-    case PinochleMelds::eightykings:
-        meldName = "eightykings";
-        break;
-    case PinochleMelds::hundredaces:
-        meldName = "hundredaces";
-        break;
-    case PinochleMelds::insuitrun:
-        meldName = "insuitrun";
-        break;
-    case PinochleMelds::doublepinochle:
-        meldName = "doublepinochle";
-        break;
-    case PinochleMelds::fourhundredjacks:
-        meldName = "fourhundredjacks";
-        break;
-    case PinochleMelds::sixhundredqueens:
-        meldName = "sixhundredqueens";
-        break;
-    case PinochleMelds::eighthundredkings:
-        meldName = "eighthundredkings";
-        break;
-    case PinochleMelds::thousandaces:
-        meldName = "thousandaces";
-        break;
-    case PinochleMelds::insuitdoublerun:
-        meldName = "insuitdoublerun";
-        break;
-    default:
-        meldName = "not_exists";
-        break;
+        case PinochleMelds::dix:
+            meldName = "dix";
+            break;
+        case PinochleMelds::offsuitmarriage:
+            meldName = "offsuitmarriage";
+            break;
+        case PinochleMelds::fortyjacks:
+            meldName = "fortyjacks";
+            break;
+        case PinochleMelds::pinochle:
+            meldName = "pinochle";
+            break;
+        case PinochleMelds::insuitmarriage:
+            meldName = "insuitmarriage";
+            break;
+        case PinochleMelds::sixtyqueens:
+            meldName = "sixtyqueens";
+            break;
+        case PinochleMelds::eightykings:
+            meldName = "eightykings";
+            break;
+        case PinochleMelds::hundredaces:
+            meldName = "hundredaces";
+            break;
+        case PinochleMelds::insuitrun:
+            meldName = "insuitrun";
+            break;
+        case PinochleMelds::doublepinochle:
+            meldName = "doublepinochle";
+            break;
+        case PinochleMelds::fourhundredjacks:
+            meldName = "fourhundredjacks";
+            break;
+        case PinochleMelds::sixhundredqueens:
+            meldName = "sixhundredqueens";
+            break;
+        case PinochleMelds::eighthundredkings:
+            meldName = "eighthundredkings";
+            break;
+        case PinochleMelds::thousandaces:
+            meldName = "thousandaces";
+            break;
+        case PinochleMelds::insuitdoublerun:
+            meldName = "insuitdoublerun";
+            break;
+        default:
+            meldName = "not_exists";
+            break;
     }
     // outputs namestring and meldscore, then return ostream
     return os << meldName << " " << PinochleGame::PinochleMeldsPointValue[static_cast<int>(pm)];
 }
 
-void PinochleGame::suit_independent_evaluation(const CardSet<PinochleRank, Suit> &cs, std::vector<PinochleMelds> &pmv)
-{
+void PinochleGame::suit_independent_evaluation(const CardSet<PinochleRank, Suit>& cs, std::vector<PinochleMelds>& pmv) {
     // call copy constructor and make an independent copy
     CardSet<PinochleRank, Suit> csl(cs);
     // data returns a pointer to the content member fo class CardSet
-    const std::vector<Card<PinochleRank, Suit>> CardSet<PinochleRank, Suit>::*pdata = CardSet<PinochleRank, Suit>::data();
+    const std::vector<Card<PinochleRank, Suit>> CardSet<PinochleRank, Suit>::* pdata = CardSet<PinochleRank, Suit>::data();
     // to use pdata, bind to an object of CardSet
     auto s = csl.*pdata;
     // sort (rank and then suit ordering relation) and reverse the vector
@@ -203,7 +195,7 @@ void PinochleGame::suit_independent_evaluation(const CardSet<PinochleRank, Suit>
             for (j = i; j < s.size() && s.at(j).rank == PinochleRank::Ace; ++j)
             {
                 count++;
-                mask |= 1 << s[j].suit;
+                mask |= 1 << static_cast<int>(s[j].suit);
             }
             if (count >= EIGHT_SAME_RANKS)
             {
@@ -220,7 +212,7 @@ void PinochleGame::suit_independent_evaluation(const CardSet<PinochleRank, Suit>
             for (j = i; j < s.size() && s.at(j).rank == PinochleRank::King; ++j)
             {
                 count++;
-                mask |= 1 << s[j].suit;
+                mask |= 1 << static_cast<int>(s[j].suit);
             }
             if (count >= EIGHT_SAME_RANKS)
             {
@@ -236,7 +228,7 @@ void PinochleGame::suit_independent_evaluation(const CardSet<PinochleRank, Suit>
             for (j = i; j < s.size() && s.at(j).rank == PinochleRank::Queen; ++j)
             {
                 count++;
-                mask |= 1 << s[j].suit;
+                mask |= 1 << static_cast<int>(s[j].suit);
                 if (s.at(j).suit == Suit::Spades)
                 {
                     spd++;
@@ -256,7 +248,7 @@ void PinochleGame::suit_independent_evaluation(const CardSet<PinochleRank, Suit>
             for (j = i; j < s.size() && s.at(j).rank == PinochleRank::Jack; ++j)
             {
                 count++;
-                mask |= 1 << s[j].suit;
+                mask |= 1 << static_cast<int>(s[j].suit);
                 if (s.at(j).suit == Suit::Diamonds)
                 {
                     dmd++;
